pyproject.toml
[project]
name = "schematized-jwt"
version = "0.1.0"
description = "A project for encoding and decoding JWT tokens using Pydantic schemas."
readme = "README.md"
authors = [
    { name = "Vladimir Boltushev", email = "vladimirboltushev@yandex.ru" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
]

[project.scripts]
schematized-jwt = "schematized_jwt:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
